+++ package.json
@@ -2,7 +2,7 @@
   "private": true,
   "name": "@live-replica/live-replica",
   "type": "module",
-  "version": "0.6.0",
+  "version": "0.6.1",
   "description": "an easy to use real-time objects replica syncing",
   "directories": {
     "test": "tests"
